move index.tsx out of helpers folder
